test: Phase4 - Knn k값 결정하는 테스트 코드 작성
